requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.35
langchain-anthropic==0.2.4
langchain-core==0.3.15
langchain-openai==0.2.5
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.5
openai==1.52.0
anthropic==0.37.1
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
pydantic-settings==2.6.0
pydantic==2.9.2
sqlparse==0.5.1
rich==13.9.4
httpx==0.27.2
